feat: base config
